chore: refactor unit test framework (#13905)
* chore: add distributed unit test

* add Clone to ClusterDescriptor

* move out ClusterDescriptor from fixture.rs

* refactor TestFixture spill out many mod: check.rs and context.rs

* make cluster nodes show work
