+++ src/query/ee/src/test_kits/context.rs
@@ -17,8 +17,7 @@ use common_exception::Result;
 use common_license::license::LicenseInfo;
 use common_meta_app::storage::StorageFsConfig;
 use common_meta_app::storage::StorageParams;
-use databend_query::test_kits::ConfigBuilder;
-use databend_query::test_kits::Setup;
+use databend_query::test_kits::*;
 use jwt_simple::algorithms::ECDSAP256KeyPairLike;
 use jwt_simple::prelude::Claims;
 use jwt_simple::prelude::Duration;
