+++ src/query/ee/tests/it/aggregating_index/index_scan.rs
@@ -28,8 +28,7 @@ use common_sql::plans::RelOperator;
 use common_sql::Planner;
 use databend_query::interpreters::InterpreterFactory;
 use databend_query::sessions::QueryContext;
-use databend_query::test_kits::fixture::expects_ok;
-use databend_query::test_kits::TestFixture;
+use databend_query::test_kits::*;
 use enterprise_query::test_kits::context::EESetup;
 use futures_util::TryStreamExt;
