+++ src/query/service/src/test_kits/fixture.rs
@@ -20,11 +20,8 @@ use common_ast::ast::Engine;
 use common_catalog::catalog_kind::CATALOG_DEFAULT;
 use common_catalog::cluster_info::Cluster;
 use common_catalog::table::AppendMode;
-use common_config::GlobalConfig;
 use common_config::InnerConfig;
-use common_config::DATABEND_COMMIT_VERSION;
 use common_exception::Result;
-use common_expression::block_debug::assert_blocks_sorted_eq_with_name;
 use common_expression::infer_table_schema;
 use common_expression::types::number::Int32Type;
 use common_expression::types::number::Int64Type;
@@ -53,21 +50,11 @@ use common_meta_app::principal::UserInfo;
 use common_meta_app::principal::UserPrivilegeSet;
 use common_meta_app::schema::DatabaseMeta;
 use common_meta_app::storage::StorageParams;
-use common_meta_types::NodeInfo;
 use common_pipeline_core::processors::ProcessorPtr;
 use common_pipeline_sinks::EmptySink;
 use common_pipeline_sources::BlocksSource;
 use common_sql::plans::CreateDatabasePlan;
 use common_sql::plans::CreateTablePlan;
-use common_sql::plans::DeletePlan;
-use common_sql::plans::UpdatePlan;
-use common_storages_fuse::FuseTable;
-use common_storages_fuse::FUSE_TBL_BLOCK_PREFIX;
-use common_storages_fuse::FUSE_TBL_LAST_SNAPSHOT_HINT;
-use common_storages_fuse::FUSE_TBL_SEGMENT_PREFIX;
-use common_storages_fuse::FUSE_TBL_SNAPSHOT_PREFIX;
-use common_storages_fuse::FUSE_TBL_SNAPSHOT_STATISTICS_PREFIX;
-use common_storages_fuse::FUSE_TBL_XOR_BLOOM_INDEX_PREFIX;
 use common_tracing::set_panic_hook;
 use futures::TryStreamExt;
 use jsonb::Number as JsonbNumber;
@@ -77,17 +64,12 @@ use log::info;
 use parking_lot::Mutex;
 use storages_common_table_meta::table::OPT_KEY_DATABASE_ID;
 use uuid::Uuid;
-use walkdir::WalkDir;
 
 use crate::clusters::ClusterDiscovery;
 use crate::clusters::ClusterHelper;
 use crate::interpreters::CreateTableInterpreter;
-use crate::interpreters::DeleteInterpreter;
 use crate::interpreters::Interpreter;
 use crate::interpreters::InterpreterFactory;
-use crate::interpreters::UpdateInterpreter;
-use crate::pipelines::executor::ExecutorSettings;
-use crate::pipelines::executor::PipelineCompleteExecutor;
 use crate::pipelines::PipelineBuildResult;
 use crate::pipelines::PipelineBuilder;
 use crate::sessions::QueryContext;
@@ -98,11 +80,13 @@ use crate::sessions::SessionType;
 use crate::sessions::TableContext;
 use crate::sql::Planner;
 use crate::storages::Table;
+use crate::test_kits::execute_pipeline;
+use crate::test_kits::ClusterDescriptor;
 use crate::test_kits::ConfigBuilder;
 use crate::GlobalServices;
 
 pub struct TestFixture {
-    default_ctx: Arc<QueryContext>,
+    pub(crate) default_ctx: Arc<QueryContext>,
     default_session: Arc<Session>,
     conf: InnerConfig,
     prefix: String,
@@ -658,6 +642,7 @@ impl TestFixture {
         )
     }
 
+    /// Generate a stream of blocks with computed columns.
     pub fn gen_computed_sample_blocks_stream(num: usize, start: i32) -> SendableDataBlockStream {
         let (_, blocks) = Self::gen_computed_sample_blocks(num, start);
         Box::pin(futures::stream::iter(blocks))
@@ -743,314 +728,4 @@ impl TestFixture {
 
 fn gen_db_name(prefix: &str) -> String {
     format!("db_{}", prefix)
-}
-
-pub fn expects_err<T>(case_name: &str, err_code: u16, res: Result<T>) {
-    if let Err(err) = res {
-        assert_eq!(
-            err.code(),
-            err_code,
-            "case name {}, unexpected error: {}",
-            case_name,
-            err
-        );
-    } else {
-        panic!(
-            "case name {}, expecting err code {}, but got ok",
-            case_name, err_code,
-        );
-    }
-}
-
-pub async fn expects_ok(
-    case_name: impl AsRef<str>,
-    res: Result<SendableDataBlockStream>,
-    expected: Vec<&str>,
-) -> Result<()> {
-    match res {
-        Ok(stream) => {
-            let blocks: Vec<DataBlock> = stream.try_collect().await?;
-            assert_blocks_sorted_eq_with_name(case_name.as_ref(), expected, &blocks)
-        }
-        Err(err) => {
-            panic!(
-                "case name {}, expecting  Ok, but got err {}",
-                case_name.as_ref(),
-                err,
-            )
-        }
-    };
-    Ok(())
-}
-
-pub async fn execute_query(ctx: Arc<QueryContext>, query: &str) -> Result<SendableDataBlockStream> {
-    let mut planner = Planner::new(ctx.clone());
-    let (plan, _) = planner.plan_sql(query).await?;
-    let executor = InterpreterFactory::get(ctx.clone(), &plan).await?;
-    executor.execute(ctx.clone()).await
-}
-
-pub fn execute_pipeline(ctx: Arc<QueryContext>, mut res: PipelineBuildResult) -> Result<()> {
-    let query_id = ctx.get_id();
-    let executor_settings = ExecutorSettings::try_create(&ctx.get_settings(), query_id)?;
-    res.set_max_threads(ctx.get_settings().get_max_threads()? as usize);
-    let mut pipelines = res.sources_pipelines;
-    pipelines.push(res.main_pipeline);
-    let executor = PipelineCompleteExecutor::from_pipelines(pipelines, executor_settings)?;
-    ctx.set_executor(executor.get_inner())?;
-    executor.execute()
-}
-
-pub async fn execute_command(ctx: Arc<QueryContext>, query: &str) -> Result<()> {
-    let res = execute_query(ctx, query).await?;
-    res.try_collect::<Vec<DataBlock>>().await?;
-    Ok(())
-}
-
-pub async fn append_sample_data(num_blocks: usize, fixture: &TestFixture) -> Result<()> {
-    append_sample_data_overwrite(num_blocks, false, fixture).await
-}
-
-pub async fn analyze_table(fixture: &TestFixture) -> Result<()> {
-    let table = fixture.latest_default_table().await?;
-    table.analyze(fixture.default_ctx.clone()).await
-}
-
-pub async fn do_deletion(ctx: Arc<QueryContext>, plan: DeletePlan) -> Result<()> {
-    let delete_interpreter = DeleteInterpreter::try_create(ctx.clone(), plan.clone())?;
-    delete_interpreter.execute(ctx).await?;
-    Ok(())
-}
-
-pub async fn do_update(ctx: Arc<QueryContext>, plan: UpdatePlan) -> Result<()> {
-    let update_interpreter = UpdateInterpreter::try_create(ctx.clone(), plan)?;
-    update_interpreter.execute(ctx).await?;
-    Ok(())
-}
-
-pub async fn append_sample_data_overwrite(
-    num_blocks: usize,
-    overwrite: bool,
-    fixture: &TestFixture,
-) -> Result<()> {
-    let stream = TestFixture::gen_sample_blocks_stream(num_blocks, 1);
-    let table = fixture.latest_default_table().await?;
-
-    let blocks = stream.try_collect().await?;
-    fixture
-        .append_commit_blocks(table.clone(), blocks, overwrite, true)
-        .await
-}
-
-pub async fn append_variant_sample_data(num_blocks: usize, fixture: &TestFixture) -> Result<()> {
-    let stream = TestFixture::gen_variant_sample_blocks_stream(num_blocks, 1);
-    let table = fixture.latest_default_table().await?;
-
-    let blocks = stream.try_collect().await?;
-    fixture
-        .append_commit_blocks(table.clone(), blocks, true, true)
-        .await
-}
-
-pub async fn append_computed_sample_data(num_blocks: usize, fixture: &TestFixture) -> Result<()> {
-    let stream = TestFixture::gen_computed_sample_blocks_stream(num_blocks, 1);
-    let table = fixture.latest_default_table().await?;
-
-    let blocks = stream.try_collect().await?;
-    fixture
-        .append_commit_blocks(table.clone(), blocks, true, true)
-        .await
-}
-
-pub async fn check_data_dir(
-    fixture: &TestFixture,
-    case_name: &str,
-    snapshot_count: u32,
-    table_statistic_count: u32,
-    segment_count: u32,
-    block_count: u32,
-    index_count: u32,
-    check_last_snapshot: Option<()>,
-    check_table_statistic_file: Option<()>,
-) -> Result<()> {
-    let data_path = match &GlobalConfig::instance().storage.params {
-        StorageParams::Fs(v) => v.root.clone(),
-        _ => panic!("storage type is not fs"),
-    };
-    let root = data_path.as_str();
-    let mut ss_count = 0;
-    let mut ts_count = 0;
-    let mut sg_count = 0;
-    let mut b_count = 0;
-    let mut i_count = 0;
-    let mut last_snapshot_loc = "".to_string();
-    let mut table_statistic_files = vec![];
-    let prefix_snapshot = FUSE_TBL_SNAPSHOT_PREFIX;
-    let prefix_snapshot_statistics = FUSE_TBL_SNAPSHOT_STATISTICS_PREFIX;
-    let prefix_segment = FUSE_TBL_SEGMENT_PREFIX;
-    let prefix_block = FUSE_TBL_BLOCK_PREFIX;
-    let prefix_index = FUSE_TBL_XOR_BLOOM_INDEX_PREFIX;
-    let prefix_last_snapshot_hint = FUSE_TBL_LAST_SNAPSHOT_HINT;
-    for entry in WalkDir::new(root) {
-        let entry = entry.unwrap();
-        if entry.file_type().is_file() {
-            let (_, entry_path) = entry.path().to_str().unwrap().split_at(root.len());
-            // trim the leading prefix, e.g. "/db_id/table_id/"
-            let path = entry_path.split('/').skip(3).collect::<Vec<_>>();
-            let path = path[0];
-            if path.starts_with(prefix_snapshot) {
-                ss_count += 1;
-            } else if path.starts_with(prefix_segment) {
-                sg_count += 1;
-            } else if path.starts_with(prefix_block) {
-                b_count += 1;
-            } else if path.starts_with(prefix_index) {
-                i_count += 1;
-            } else if path.starts_with(prefix_snapshot_statistics) {
-                ts_count += 1;
-                table_statistic_files.push(entry_path.to_string());
-            } else if path.starts_with(prefix_last_snapshot_hint) && check_last_snapshot.is_some() {
-                let content = fixture
-                    .default_ctx
-                    .get_data_operator()?
-                    .operator()
-                    .read(entry_path)
-                    .await?;
-                last_snapshot_loc = str::from_utf8(&content)?.to_string();
-            }
-        }
-    }
-
-    assert_eq!(
-        ss_count, snapshot_count,
-        "case [{}], check snapshot count",
-        case_name
-    );
-    assert_eq!(
-        ts_count, table_statistic_count,
-        "case [{}], check snapshot statistics count",
-        case_name
-    );
-    assert_eq!(
-        sg_count, segment_count,
-        "case [{}], check segment count",
-        case_name
-    );
-
-    assert_eq!(
-        b_count, block_count,
-        "case [{}], check block count",
-        case_name
-    );
-
-    assert_eq!(
-        i_count, index_count,
-        "case [{}], check index count",
-        case_name
-    );
-
-    if check_last_snapshot.is_some() {
-        let table = fixture.latest_default_table().await?;
-        let fuse_table = FuseTable::try_from_table(table.as_ref())?;
-        let snapshot_loc = fuse_table.snapshot_loc().await?;
-        let snapshot_loc = snapshot_loc.unwrap();
-        assert!(last_snapshot_loc.contains(&snapshot_loc));
-        assert_eq!(
-            last_snapshot_loc.find(&snapshot_loc),
-            Some(last_snapshot_loc.len() - snapshot_loc.len())
-        );
-    }
-
-    if check_table_statistic_file.is_some() {
-        let table = fixture.latest_default_table().await?;
-        let fuse_table = FuseTable::try_from_table(table.as_ref())?;
-        let snapshot_opt = fuse_table.read_table_snapshot().await?;
-        assert!(snapshot_opt.is_some());
-        let snapshot = snapshot_opt.unwrap();
-        let ts_location_opt = snapshot.table_statistics_location.clone();
-        assert!(ts_location_opt.is_some());
-        let ts_location = ts_location_opt.unwrap();
-        println!(
-            "ts_location_opt: {:?}, table_statistic_files: {:?}",
-            ts_location, table_statistic_files
-        );
-        assert!(
-            table_statistic_files
-                .iter()
-                .any(|e| e.contains(&ts_location))
-        );
-    }
-
-    Ok(())
-}
-
-pub async fn history_should_have_item(
-    fixture: &TestFixture,
-    case_name: &str,
-    item_cnt: u32,
-) -> Result<()> {
-    // check history
-    let db = fixture.default_db_name();
-    let tbl = fixture.default_table_name();
-    let count_str = format!("| {}        |", item_cnt);
-    let expected = vec![
-        "+----------+",
-        "| Column 0 |",
-        "+----------+",
-        count_str.as_str(),
-        "+----------+",
-    ];
-
-    let qry = format!(
-        "select count(*) as count from fuse_snapshot('{}', '{}')",
-        db, tbl
-    );
-
-    expects_ok(
-        format!("{}: count_of_history_item_should_be_1", case_name),
-        fixture.execute_query(qry.as_str()).await,
-        expected,
-    )
-    .await
-}
-
-pub struct ClusterDescriptor {
-    local_node_id: String,
-    cluster_nodes_list: Vec<Arc<NodeInfo>>,
-}
-
-impl ClusterDescriptor {
-    pub fn new() -> ClusterDescriptor {
-        ClusterDescriptor {
-            local_node_id: String::from(""),
-            cluster_nodes_list: vec![],
-        }
-    }
-
-    pub fn with_node(self, id: impl Into<String>, addr: impl Into<String>) -> ClusterDescriptor {
-        let mut new_nodes = self.cluster_nodes_list.clone();
-        new_nodes.push(Arc::new(NodeInfo::create(
-            id.into(),
-            0,
-            addr.into(),
-            DATABEND_COMMIT_VERSION.to_string(),
-        )));
-        ClusterDescriptor {
-            cluster_nodes_list: new_nodes,
-            local_node_id: self.local_node_id,
-        }
-    }
-
-    pub fn with_local_id(self, id: impl Into<String>) -> ClusterDescriptor {
-        ClusterDescriptor {
-            local_node_id: id.into(),
-            cluster_nodes_list: self.cluster_nodes_list,
-        }
-    }
-}
-
-impl Default for ClusterDescriptor {
-    fn default() -> Self {
-        Self::new()
-    }
 }
