+++ src/query/service/tests/it/api/http/logs.rs
@@ -15,7 +15,7 @@
 use common_base::base::tokio;
 use common_exception::Result;
 use databend_query::api::http::v1::logs::logs_handler;
-use databend_query::test_kits::TestFixture;
+use databend_query::test_kits::*;
 use poem::get;
 use poem::http::Method;
 use poem::http::StatusCode;
