+++ src/query/service/tests/it/auth.rs
@@ -24,8 +24,7 @@ use common_users::UserApiProvider;
 use databend_query::auth::AuthMgr;
 use databend_query::auth::Credential;
 use databend_query::sessions::TableContext;
-use databend_query::test_kits::ConfigBuilder;
-use databend_query::test_kits::TestFixture;
+use databend_query::test_kits::*;
 use jwt_simple::prelude::*;
 use p256::EncodedPoint;
 use wiremock::matchers::method;
