+++ src/query/service/tests/it/clusters.rs
@@ -16,8 +16,7 @@ use common_base::base::tokio;
 use common_exception::Result;
 use databend_query::clusters::ClusterDiscovery;
 use databend_query::clusters::ClusterHelper;
-use databend_query::test_kits::ConfigBuilder;
-use databend_query::test_kits::TestFixture;
+use databend_query::test_kits::*;
 use pretty_assertions::assert_eq;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
