+++ src/query/service/tests/it/databases/system/system_database.rs
@@ -17,7 +17,7 @@ use databend_query::catalogs::InMemoryMetas;
 use databend_query::catalogs::SYS_DB_ID_BEGIN;
 use databend_query::catalogs::SYS_TBL_ID_BEGIN;
 use databend_query::databases::SystemDatabase;
-use databend_query::test_kits::ConfigBuilder;
+use databend_query::test_kits::*;
 
 #[test]
 fn test_disable_system_table() -> Result<()> {
