+++ src/query/service/tests/it/distributed/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2021 Datafuse Labs
+// Copyright 2021 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,20 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#![allow(clippy::too_many_arguments)]
-
-mod block_writer;
-mod check;
 mod cluster;
-pub mod config;
-mod context;
-mod fixture;
-mod fuse;
-
-pub use block_writer::BlockWriter;
-pub use check::*;
-pub use cluster::ClusterDescriptor;
-pub use config::ConfigBuilder;
-pub use context::*;
-pub use fixture::*;
-pub use fuse::*;
