+++ src/query/service/tests/it/storages/fuse/operations/commit.rs
@@ -112,7 +112,7 @@ use common_storages_fuse::FUSE_TBL_SNAPSHOT_PREFIX;
 use common_users::GrantObjectVisibilityChecker;
 use dashmap::DashMap;
 use databend_query::sessions::QueryContext;
-use databend_query::test_kits::fixture::TestFixture;
+use databend_query::test_kits::*;
 use futures::TryStreamExt;
 use parking_lot::RwLock;
 use storages_common_table_meta::meta::Location;
