+++ src/query/service/tests/it/storages/fuse/operations/gc.rs
@@ -21,12 +21,7 @@ use common_catalog::table_context::TableContext;
 use common_exception::Result;
 use common_storages_fuse::io::MetaWriter;
 use common_storages_fuse::FuseTable;
-use databend_query::test_kits::fixture::append_sample_data;
-use databend_query::test_kits::fixture::check_data_dir;
-use databend_query::test_kits::fixture::TestFixture;
-use databend_query::test_kits::utils::generate_segments;
-use databend_query::test_kits::utils::generate_snapshot_with_segments;
-use databend_query::test_kits::utils::generate_snapshots;
+use databend_query::test_kits::*;
 use storages_common_table_meta::meta::Location;
 use storages_common_table_meta::meta::TableSnapshot;
 use storages_common_table_meta::meta::Versioned;
