+++ src/query/service/tests/it/storages/fuse/operations/mutation/block_compact_mutator.rs
@@ -30,8 +30,7 @@ use databend_query::pipelines::executor::PipelineCompleteExecutor;
 use databend_query::schedulers::build_query_pipeline_without_render_result_set;
 use databend_query::sessions::QueryContext;
 use databend_query::sessions::TableContext;
-use databend_query::test_kits::fixture::expects_ok;
-use databend_query::test_kits::fixture::TestFixture;
+use databend_query::test_kits::*;
 use rand::thread_rng;
 use rand::Rng;
 use storages_common_table_meta::meta::SegmentInfo;
