+++ src/query/service/tests/it/storages/fuse/operations/purge_drop.rs
@@ -14,9 +14,7 @@
 
 use common_base::base::tokio;
 use common_exception::Result;
-use databend_query::test_kits::fixture::append_sample_data;
-use databend_query::test_kits::fixture::check_data_dir;
-use databend_query::test_kits::fixture::TestFixture;
+use databend_query::test_kits::*;
 
 #[tokio::test(flavor = "multi_thread")]
 async fn test_fuse_snapshot_truncate_in_drop_stmt() -> Result<()> {
