+++ src/query/service/tests/it/storages/fuse/operations/table_analyze.rs
@@ -27,12 +27,7 @@ use databend_query::sessions::QueryContext;
 use databend_query::sessions::TableContext;
 use databend_query::sql::plans::Plan;
 use databend_query::sql::Planner;
-use databend_query::test_kits::fixture::analyze_table;
-use databend_query::test_kits::fixture::do_deletion;
-use databend_query::test_kits::fixture::do_update;
-use databend_query::test_kits::fixture::execute_command;
-use databend_query::test_kits::fixture::TestFixture;
-use databend_query::test_kits::utils::query_count;
+use databend_query::test_kits::*;
 use storages_common_cache::LoadParams;
 use storages_common_table_meta::meta::SegmentInfo;
 use storages_common_table_meta::meta::Statistics;
