+++ src/query/service/tests/it/storages/fuse/pruning.rs
@@ -42,7 +42,7 @@ use databend_query::sessions::TableContext;
 use databend_query::storages::fuse::io::MetaReaders;
 use databend_query::storages::fuse::FUSE_OPT_KEY_BLOCK_PER_SEGMENT;
 use databend_query::storages::fuse::FUSE_OPT_KEY_ROW_PER_BLOCK;
-use databend_query::test_kits::fixture::TestFixture;
+use databend_query::test_kits::*;
 use opendal::Operator;
 use storages_common_cache::LoadParams;
 use storages_common_table_meta::meta::BlockMeta;
