+++ src/query/service/tests/it/storages/fuse/table.rs
@@ -20,7 +20,7 @@ use common_meta_app::schema::TableInfo;
 use common_sql::executor::table_read_plan::ToReadDataSourcePlan;
 use databend_query::storages::fuse::FuseTable;
 use databend_query::stream::ReadDataBlockStream;
-use databend_query::test_kits::fixture::TestFixture;
+use databend_query::test_kits::*;
 use futures::TryStreamExt;
 use storages_common_table_meta::table::OPT_KEY_DATABASE_ID;
